Add files via upload
This project aims to develop a deep learning model to predict the age of individuals based on facial images, utilizing the UTKFace dataset with over 20,000 annotated images. The main objective is to explore and compare different neural network architectures, such as MobileNet and Vision Outlooker (VOLO), to identify the most accurate model for age prediction. Additionally, Class Activation Maps (CAM) are applied to the models to provide interpretability and explanations for the model’s decision-making process.